refactor(e2e): split tests into separate files
