+++ package.json
@@ -26,7 +26,7 @@
     "make-docs": "./node_modules/.bin/jsdoc index.js lib -r -R README.md -u ./doc -c .jsdoc.json --verbose -d ./jsdoc",
     "publish-docs": "gh-pages -d jsdoc --repo git@github.com:internxt/bridge.git",
     "build": "tsc",
-    "test": "jest --testPathIgnorePatterns ./tests/lib/mongo",
+    "test": "jest --testPathIgnorePatterns ./tests/lib/mongo ./tests/lib/e2e",
     "test:cov": "jest --coverage",
     "test:e2e": "NODE_ENV=test jest --config ./tests/lib/e2e/jest-e2e.json --setupFiles dotenv/config",
     "test-mongo-init": "ts-node ./tests/lib/mongo/init",
