+++ tests/lib/e2e/global-teardown.ts
@@ -1,0 +1,3 @@
+export default async () => {
+  process.emit('SIGINT')
+}
